freebsd/rtemsbsd: Resolve mismatched type errors

+++ testsuite/usbkbd01/init.c
@@ -65,7 +65,7 @@ usb_keyboard_read_task(rtems_task_argument arg)
 {
 	rtems_status_code sc;
 	struct usb_test_message msg;
-	uint32_t size;
+	size_t size;
 	int bytes;
 
 	rtask_active = true;
@@ -109,7 +109,7 @@ usb_keyboard_open_task(rtems_task_argument arg)
 {
 	rtems_status_code sc;
 	struct usb_test_message msg;
-	uint32_t size;
+	size_t size;
 	int fd;
 
 	fd = -2;
